ptdump: support perf_event sideband
Add a new options

  --pevent <file>[:<from>[-<to>]]
  --pevent:primary <file>[:<from>[-<to>]]
  --pevent:secondary <file>[:<from>[-<to>]]

to dump the Linux perf_event sideband contained in <file>.  An optional offset
or range can be given as <from> and <to> file offsets.  The sideband records
are interleaved with the trace packets based on TSC.

Add new options

  --pevent:sample-type
  --pevent:time-zero
  --pevent:time-shift
  --pevent:time-mult
  --pevent:tsc-offset

to configure sideband iterpretation and new options

  --pevent:show-filename
  --pevent:verbose

to control the output.

We also ignore some options that are used by ptxed but that are not relevant for
ptdump:

  --pevent:kernel-start
  --pevent:kcore
  --pevent:ring-0
  --pevent:ring-3

This feature requires the PEVENT cmake option to be ON.

Change-Id: I4bead56618b7c480eb175352a0fef9d4b4cd2da4
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>